harden: C++ solver overflow/tolerance guards + statistical coverage tests
C++ solver hardening (#13), guards that only affect extreme-scale inputs:
- gabow_tarjan: reject cost ranges that overflow the (n+1) bit-scaling product
  or collide with the 1e15 forbidden sentinel (silently-wrong optimal).
- 64-bit flat indexing in prepare_cost_matrix, solve_sinkhorn, and the auction
  eps (n*m and nn*nn wrapped past ~2^31).
- network_simplex and lapmod reject problems whose arc / entry counts exceed a
  32-bit index.
- solve_munkres: magnitude-scaled zero tolerance (fixed 1e-12 made a solvable
  large-cost matrix throw after reductions).
- full_matching: report infeasible when group capacity < number of units
  instead of dropping units as "optimal".
- morph_pixel_level internal helpers assert pixel/assignment buffer sizes.
- cosmetic: fix duplicated n==0 guards (bruteforce/ssp/hk01 rcpp), correct the
  build_allowed mask-convention comment (0 = allowed), drop the stale csa
  "padded to square" comment.

Statistical coverage tests (#14): add known-effect recovery across seeds
(matched vs naive bias) and matched-pair 95% CI coverage (>= 85% over 30
seeds) to test-statistical-recovery.R. Verified empirically: matched bias
0.03 vs naive 0.50; coverage 30/30.

Full test suite passes with NOT_CRAN=true (5129 tests, 0 failures) after
recompiling.

Fixes #14
Progress toward #13

+++ src/solvers/solve_auction_rcpp.cpp
@@ -252,7 +252,8 @@ Rcpp::List solve_auction_scaled_impl(Rcpp::NumericMatrix cost, bool maximize,
 
   double epsilon = std::max(1.0, max_abs_cost * initial_epsilon_factor);
   // Use smaller final epsilon for better optimality guarantee
-  double eps_final = (final_epsilon > 0.0) ? final_epsilon : std::min(1e-6, 1.0 / (nn * nn));
+  double eps_final = (final_epsilon > 0.0) ? final_epsilon
+                     : std::min(1e-6, 1.0 / (static_cast<double>(nn) * nn));
 
   // State: prices persist, matching rebuilt each phase
   std::vector<double> price(m, 0.0);
